+++ python_src/models/ac3rp/__init__.py
@@ -2,3 +2,4 @@ from .road import Road
 from .vehicle import Vehicle
 from .scenario import CrashScenario
 from .report import Report
+from .movement import Movement
